refactor: separate ThemeContext and ThemeProvider for better Fast Refresh
