+++ rs-react-app/src/components/SearchBar.tsx
