+++ rs-react-app/src/contexts/ThemeProvider.tsx
@@ -1,0 +1,24 @@
+import { useState, type ReactNode } from 'react';
+import { ThemeContext } from './ThemeContext';
+
+type Theme = 'light' | 'dark';
+
+interface ThemeProviderProps {
+  children: ReactNode;
+}
+
+export const ThemeProvider = ({ children }: ThemeProviderProps) => {
+  const [theme, setThemeState] = useState<Theme>('light');
+
+  const toggleTheme = () => {
+    setThemeState((prev) => (prev === 'light' ? 'dark' : 'light'));
+  };
+
+  return (
+    <ThemeContext.Provider
+      value={{ theme, toggleTheme, setTheme: setThemeState }}
+    >
+      {children}
+    </ThemeContext.Provider>
+  );
+};
