+++ rs-react-app/src/hooks/useTheme.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { ThemeContext } from './ThemeContextContext';
+import { ThemeContext } from '../contexts/ThemeContext';
 
 export const useTheme = () => {
   const context = useContext(ThemeContext);
