+++ rs-react-app/src/main.tsx
@@ -4,7 +4,7 @@ import './index.css';
 import App from './App.tsx';
 import ErrorBoundary from './components/ErrorBoundary.tsx';
 import { BrowserRouter } from 'react-router-dom';
-import { ThemeProvider } from './contexts/ThemeContext';
+import { ThemeProvider } from './contexts/ThemeProvider';
 
 const rootEl = document.getElementById('root');
 if (rootEl) {
